+++ Core/src/gfx/Render/Jobs/InstancedJob.cpp
@@ -14,7 +14,7 @@ void tryn::gfx::InstancedJob::Execute(IGraphics& gfx)
 	data.pDrawable->BindExtraBinds();
 	data.pDrawable->BindTransformCBuf();
 	data.pStep->Bind();
-	gfx.DrawIndexedInstanced(data.pDrawable->GetIndexCount(), data.transforms.size(), 0u, 0u, 0u);
+	gfx.DrawIndexedInstanced((int)data.pDrawable->GetIndexCount(), (int)data.transforms.size(), 0u, 0u, 0u);
 }
 
 void tryn::gfx::InstancedJob::Execute(IContext& ctx)
@@ -23,5 +23,5 @@ void tryn::gfx::InstancedJob::Execute(IContext& ctx)
 	data.pDrawable->BindExtraBinds(ctx);
 	data.pDrawable->BindTransformCBuf(ctx);
 	data.pStep->Bind(ctx);
-	ctx.DrawIndexedInstanced(data.pDrawable->GetIndexCount(), data.transforms.size(), 0u, 0u, 0u);
+	ctx.DrawIndexedInstanced((int)data.pDrawable->GetIndexCount(), (int)data.transforms.size(), 0u, 0u, 0u);
 }
