+++ Core/src/mem/ArenaAllocator.h
@@ -25,7 +25,7 @@ namespace tryn::mem
 			}
 
 			auto& ref = buffer[usedChunks];
-			usedChunks += numChunks;
+			usedChunks += (uint32_t)numChunks;
 
 			return reinterpret_cast<void*>(&ref);
 		}
@@ -38,7 +38,7 @@ namespace tryn::mem
 		template <typename T, typename... Args>
 		std::span<T> MakeNewArray(std::size_t numElements, Args... args)
 		{
-			auto ptr = static_cast<T*>(Allocate(sizeof(T) * numElements));
+			auto ptr = static_cast<T*>(Allocate((uint32_t)(sizeof(T) * numElements)));
 			const auto start = new(ptr) T{ args... };
 			for (auto i = 1; i < numElements; i++)
 			{
@@ -50,7 +50,7 @@ namespace tryn::mem
 		{
 			if (overflow)
 			{
-				buffer.resize(buffer.size() * 1.3f);
+				buffer.resize((size_t)(buffer.size() * 1.3f));
 				overflow = false;
 			}
 			std::memset(buffer.data(), 0, buffer.size());
